+++ camera/v3/fake-pipeline2/OMXDecoder.cpp
@@ -458,7 +458,7 @@ bool OMXDecoder::uvm_buffer_init() {
         eRet = OMX_UseBuffer(mVDecoderHandle,
                     &bufferHdr,
                     mVideoOutputPortParam.nPortIndex,
-                    (OMX_PTR)(long)shared_fd,
+                    (OMX_PTR)shared_fd,
                     mVideoOutputPortParam.nBufferSize,
                     (OMX_U8*)0xFFFF);//(OMX_U8*)cpu_ptr);
         if (OMX_ErrorNone != eRet) {
@@ -658,14 +658,14 @@ bool OMXDecoder::ion_buffer_init() {
         eRet = OMX_UseBuffer(mVDecoderHandle,
                 &bufferHdr,
                 mVideoOutputPortParam.nPortIndex,
-                (OMX_PTR)(long)shared_fd,
+                (OMX_PTR)shared_fd,
                 mVideoOutputPortParam.nBufferSize,
                 (OMX_U8*)cpu_ptr);
         if (OMX_ErrorNone != eRet) {
             ALOGE("OMX_UseBuffer on output port failed! eRet = %#x\n", eRet);
             return false;
         }
-        bufferHdr->pAppPrivate = (OMX_PTR)(long)ion_hnd;
+        bufferHdr->pAppPrivate = (OMX_PTR)ion_hnd;
         mListOfOutputBufferHeader.push_back(bufferHdr);
     }
     return true;
@@ -681,7 +681,7 @@ bool OMXDecoder::do_buffer_init() {
         eRet = OMX_UseBuffer(mVDecoderHandle,
                 &bufferHdr,
                 mVideoOutputPortParam.nPortIndex,
-                (OMX_PTR)(long)shared_fd,
+                (OMX_PTR)shared_fd,
                 mVideoOutputPortParam.nBufferSize,
                 (OMX_U8*)cpu_ptr);
         if (OMX_ErrorNone != eRet) {
@@ -807,12 +807,12 @@ void OMXDecoder::free_ion_buffer(void) {
         if (bufferHdr != NULL) {
             if (mUseDMABuffer) {
                 munmap(bufferHdr->pBuffer, mWidth * mHeight * 3 / 2);
-                int ret = close((int)(long)bufferHdr->pPlatformPrivate);
+                int ret = close((int)bufferHdr->pPlatformPrivate);
                 if (ret != 0) {
                     ALOGD("close ion shared fd failed for reason %s",strerror(errno));
                 }
                 ALOGD("bufferHdr->pAppPrivate: %p", bufferHdr->pAppPrivate);
-                ret = ion_free(mIonFd, (ion_user_handle_t)(long)(bufferHdr->pAppPrivate));
+                ret = ion_free(mIonFd, (ion_user_handle_t)(bufferHdr->pAppPrivate));
                 if (ret != 0) {
                     ALOGD("ion_free failed for reason %s",strerror(errno));
                 }
@@ -858,11 +858,11 @@ void OMXDecoder::free_uvm_buffer() {
            OMX_ERRORTYPE err;
            if (bufferHdr != NULL) {
                if (mUseDMABuffer) {
-                   LOG_LINE("try to unmap uvm vaddr %p, fd: %d", bufferHdr->pBuffer, (int)(long)(bufferHdr->pPlatformPrivate));
+                   LOG_LINE("try to unmap uvm vaddr %p, fd: %d", bufferHdr->pBuffer, (int)(bufferHdr->pPlatformPrivate));
 
                    //munmap(bufferHdr->pBuffer, mWidth * mHeight * 3 / 2);
 
-                   amuvm_free((int)(long)(bufferHdr->pPlatformPrivate));
+                   amuvm_free((int)(bufferHdr->pPlatformPrivate));
 
                    ALOGD("bufferHdr->pAppPrivate: %p", bufferHdr->pAppPrivate);
 
@@ -1070,15 +1070,15 @@ void OMXDecoder::SetOutputBuffer(int share_fd, uint8_t* addr) {
 
     if (!mListOfOutputBufferHeader.empty()) {
         pBufferHdr = *mListOfOutputBufferHeader.begin();
-        pBufferHdr->pAppPrivate = (void *)(long)share_fd;
-        pBufferHdr->pPlatformPrivate = (void *)(long)share_fd;
+        pBufferHdr->pAppPrivate = (void *)share_fd;
+        pBufferHdr->pPlatformPrivate = (void *)share_fd;
         pBufferHdr->pBuffer = addr;
 
         ALOGV("SetOutputBuffer %p, OMX_FillThisBuffer, share_fd=%d, pAppPrivate=%d, pPlatformPrivate=%d",
                 pBufferHdr,
                 share_fd,
-                (int)(long)(pBufferHdr->pAppPrivate),
-                (int)(long)(pBufferHdr->pPlatformPrivate));
+                (int)(pBufferHdr->pAppPrivate),
+                (int)(pBufferHdr->pPlatformPrivate));
 
         OMX_FillThisBuffer(mVDecoderHandle, pBufferHdr);
         ALOGV("SetOutputBuffer: erase mListOfOutputBufferHeader");
@@ -1102,7 +1102,7 @@ int OMXDecoder::DequeueBuffer(int dst_fd ,uint8_t* dst_buf,
 #ifdef GE2D_ENABLE
            if (dst_fd != -1) {
                 //copy data using ge2d
-                int omx_share_fd = (int)(long)pOutPutBufferHdr->pPlatformPrivate;
+                int omx_share_fd = (int)pOutPutBufferHdr->pPlatformPrivate;
                 if (mGE2D) {
                     mGE2D->ge2d_copy(dst_fd,omx_share_fd,dst_w,dst_h,ge2dTransform::NV12);
                 }else
@@ -1110,7 +1110,7 @@ int OMXDecoder::DequeueBuffer(int dst_fd ,uint8_t* dst_buf,
             }
             else {
                 if (mem_type == UVM_BUFFER) {
-                    uint8_t* cpu_ptr = (uint8_t*)mmap(NULL, pOutPutBufferHdr->nFilledLen, PROT_READ | PROT_WRITE, MAP_SHARED, (int)(long)pOutPutBufferHdr->pPlatformPrivate, 0);
+                    uint8_t* cpu_ptr = (uint8_t*)mmap(NULL, pOutPutBufferHdr->nFilledLen, PROT_READ | PROT_WRITE, MAP_SHARED, (int)pOutPutBufferHdr->pPlatformPrivate, 0);
                     memcpy(dst_buf, cpu_ptr, pOutPutBufferHdr->nFilledLen);
                     munmap(cpu_ptr, pOutPutBufferHdr->nFilledLen);
                 } else
